# KAFKA-8904: Improve producer's topic metadata fetching.