… clients/src/main/java/org/apache/kafka/clients/producer/KafkaProducer.java
… clients/src/main/java/org/apache/kafka/clients/producer/ProducerConfig.java
… clients/src/main/java/org/apache/kafka/clients/producer/internals/ProducerMetadata.java
… clients/src/test/java/org/apache/kafka/clients/MetadataTest.java
… clients/src/test/java/org/apache/kafka/clients/MockClient.java
… clients/src/test/java/org/apache/kafka/clients/NetworkClientTest.java
… clients/src/test/java/org/apache/kafka/clients/consumer/internals/ConsumerCoordinatorTest.java
… clients/src/test/java/org/apache/kafka/clients/consumer/internals/ConsumerMetadataTest.java
… clients/src/test/java/org/apache/kafka/clients/consumer/internals/ConsumerNetworkClientTest.java
… clients/src/test/java/org/apache/kafka/clients/consumer/internals/FetcherTest.java
… clients/src/test/java/org/apache/kafka/clients/producer/KafkaProducerTest.java
… clients/src/test/java/org/apache/kafka/clients/producer/internals/ProducerMetadataTest.java
… clients/src/test/java/org/apache/kafka/clients/producer/internals/SenderTest.java
… clients/src/test/java/org/apache/kafka/clients/producer/internals/TransactionManagerTest.java