Improves the producer's topic metadata fetching logic by maintaining a per-topic last refresh time, which is used to discriminate which topics' metadata should be fetched.

The improvement can be witnessed during producer startup, where many topics may be fetched serially. Previously, where N topics were being fetched serially, O(N^2) total topic metadata would be processed, whereas this improvement makes it O(N).

A simple test can show the differences fairly easily. Starting a producer on 500 existing topics with 64 partitions each, single-threaded:

Old: 1000 records sent, 32.446463 records/sec (0.00 MB/sec), 65.52 ms avg latency, 459.00 ms max latency, 23 ms 50th, 213 ms 95th, 360 ms 99th, 459 ms 99.9th.

New: 1000 records sent, 206.143063 records/sec (0.00 MB/sec), 10.55 ms avg latency, 187.00 ms max latency, 7 ms 50th, 29 ms 95th, 43 ms 99th, 187 ms 99.9th.

### Committer Checklist (excluded from commit message)
- [ ] Verify design and implementation 
- [ ] Verify test coverage and CI build status
- [ ] Verify documentation (including upgrade notes)