@bdbyrne 2019-12-13
One unforeseen issue is that some tests expect the Metadata to discover when topics have been deleted because they issue requests for all topic metadata and then rely on the cache to be rebuilt, thus discarding any topics that may have been deleted since the last update. With the merge code, the old metadata will be retained if it's still wanted, e.g. it matches the consumer's subscription pattern. Easy but non-elegant fix, looking for something cleaner.