@ijuma 2020-01-29
> Old: 1000 records sent, 32.446463 records/sec (0.00 MB/sec), 65.52 ms avg latency, 459.00 ms max latency, 23 ms 50th, 213 ms 95th, 360 ms 99th, 459 ms 99.9th.
> 
> New: 1000 records sent, 206.143063 records/sec (0.00 MB/sec), 10.55 ms avg latency, 187.00 ms max latency, 7 ms 50th, 29 ms 95th, 43 ms 99th, 187 ms 99.9th.

Awesome!