@bdbyrne 2020-01-30
> Btw, the 0.00 MB/s bit is wrong, right?

I was using tiny records since it's more about evaluating how fast the initial metadata is fetched. Both scenarios eventually have the metadata cached, and once the buffers start filling they're able to hit max throughput.

Old:
423 records sent, 84.1 records/sec (0.04 MB/sec), 22.0 ms avg latency, 157.0 ms max latency.
393 records sent, 77.9 records/sec (0.04 MB/sec), 27.2 ms avg latency, 119.0 ms max latency.
764 records sent, 152.0 records/sec (0.07 MB/sec), 21.6 ms avg latency, 140.0 ms max latency.
13077 records sent, 2615.4 records/sec (1.28 MB/sec), 19.1 ms avg latency, 259.0 ms max latency.
9681 records sent, 1936.2 records/sec (0.95 MB/sec), 7.2 ms avg latency, 19.0 ms max latency.
9091 records sent, 1818.2 records/sec (0.89 MB/sec), 8.0 ms avg latency, 27.0 ms max latency.
9947 records sent, 1989.0 records/sec (0.97 MB/sec), 8.6 ms avg latency, 25.0 ms max latency.
13094 records sent, 2617.8 records/sec (1.28 MB/sec), 10.0 ms avg latency, 35.0 ms max latency.
23459 records sent, 4690.9 records/sec (2.29 MB/sec), 12.2 ms avg latency, 31.0 ms max latency.
342941 records sent, 68588.2 records/sec (33.49 MB/sec), 248.4 ms avg latency, 1577.0 ms max latency.
952536 records sent, 190393.0 records/sec (92.97 MB/sec), 290.9 ms avg latency, 1866.0 ms max latency.
1008257 records sent, 201490.2 records/sec (98.38 MB/sec), 277.4 ms avg latency, 2268.0 ms max latency.
1008358 records sent, 201108.5 records/sec (98.20 MB/sec), 274.9 ms avg latency, 2075.0 ms max latency.
1093013 records sent, 218602.6 records/sec (106.74 MB/sec), 253.4 ms avg latency, 1681.0 ms max latency.

New:
8275 records sent, 1655.0 records/sec (0.81 MB/sec), 3.5 ms avg latency, 158.0 ms max latency.
12695 records sent, 2538.5 records/sec (1.24 MB/sec), 5.1 ms avg latency, 20.0 ms max latency.
9115 records sent, 1822.3 records/sec (0.89 MB/sec), 8.7 ms avg latency, 62.0 ms max latency.
10442 records sent, 2087.6 records/sec (1.02 MB/sec), 8.0 ms avg latency, 25.0 ms max latency.
13318 records sent, 2663.1 records/sec (1.30 MB/sec), 8.8 ms avg latency, 30.0 ms max latency.
21180 records sent, 4234.3 records/sec (2.07 MB/sec), 11.0 ms avg latency, 53.0 ms max latency.
255753 records sent, 51109.7 records/sec (24.96 MB/sec), 178.9 ms avg latency, 939.0 ms max latency.
1052303 records sent, 210250.3 records/sec (102.66 MB/sec), 265.4 ms avg latency, 1791.0 ms max latency.
1080308 records sent, 216018.4 records/sec (105.48 MB/sec), 257.4 ms avg latency, 1567.0 ms max latency.
1056188 records sent, 211153.1 records/sec (103.10 MB/sec), 259.6 ms avg latency, 1671.0 ms max latency.
1098043 records sent, 219608.6 records/sec (107.23 MB/sec), 254.9 ms avg latency, 1481.0 ms max latency.

I stress-tested a 10,000 single-partition topics setup to verify:

Old:
real	10m17.375s
user	7m31.150s
sys	0m2.766s

New:
real	0m40.362s
user	1m0.020s
sys	0m0.961s

The average latency was about 5-10x in the old, and max latency about 20-30x, which I wouldn't expect if it's just the metadata response payload size that's increased. In other words, there may be some improvements to be found on the broker, needs further profiling to see where the time is going.