@bdbyrne 2020-02-07
> Looks like this PR needs rebasing.

Done. New metadata handling in MetadataCache simplified the node resolution code for me.